New failing test: After this test is fixed to work as intended, it shows that the Circle is still not producing closed circles
The previous test syntax was not working as intended because it was not
taking into account the nesting of the coordinates structure.

If you simply dump out the entire coordinate list returned, you can see
that the last point is not the same as the first point, as it should be.

This test update correctly takes the nesting into account and tests
both the latitude and longitude.